allauthdemo/requirements.txt

Django==2.2.10
django-allauth==0.42.0
django-contrib-comments==1.9.2
django-crispy-forms==1.9.2
django-haystack==2.8.1
django-markdown-deux==1.0.5
django-markdownx==3.0.1
django-taggit==1.3.0
gevent==1.3.6
httpie==2.2.0
Markdown==3.2.2
markdown2==2.3.9
oauthlib==3.1.0
pandas==0.23.4
Pillow==5.2.0
pip==20.2.2
Pygments==2.6.1
PyMySQL==0.9.3
six==1.11.0
urllib3==1.23
elasticsearch==2.4.1
